new FootGroundContactParameters (double values are accepted now, FootGroundContactDetector.h) improved striker, decide_and_go_to_ready_position (wait a short time even the robot is localized), play_soccer (look at ball if in game state set) try to clean up search_for_ball

+++ NaoTHSoccer/Source/Core/Motion/MorphologyProcessor/FootGroundContactDetector.h
@@ -60,9 +60,9 @@ private:
       DebugParameterList::getInstance().remove(this);
     }
 
-    int left;
-    int right;
-    int invalid;
+    double left;
+    double right;
+    double invalid;
   };
 
   Parameters footParams;
